refactor(desktop): move streaming preview helper to its own module
`StreamingFileWrite.tsx` exported `buildStreamingFilePreview` beside the
component. React Fast Refresh cannot preserve component state when a
component file also exports non-components.

Move the helper and its `StreamingFilePreview` type into
`streaming-file-preview.ts`. The component file exports only the component.
